third_party/blink/web_tests/svg/custom/poly-parsing-error-expected.txt

CONSOLE ERROR: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "…0,250 80,200 250".
CONSOLE ERROR: Error: <polyline> attribute points: Unexpected end of attribute. Expected number, "…,250 180,200 250".
CONSOLE ERROR: Error: <polygon> attribute points: Unexpected end of attribute. Expected number, "…0,160 150,110 80".
CONSOLE ERROR: Error: <polygon> attribute points: Unexpected end of attribute. Expected number, "…,160 250,110 180".
Tests whether polygons render up to first parsing error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS poly1.points.numberOfItems is 4
PASS poly2.points.numberOfItems is 4
PASS poly3.points.numberOfItems is 4
PASS poly4.points.numberOfItems is 4
PASS poly5.points.numberOfItems is 3
PASS poly6.points.numberOfItems is 3
PASS poly7.points.numberOfItems is 3
PASS poly8.points.numberOfItems is 3
PASS successfullyParsed is true

TEST COMPLETE
